+++ Editor/SVGImporter.cs
@@ -324,7 +324,10 @@ namespace Unity.VectorGraphics.Editor
 
             var tex = BuildTexture(sprite, name);
             var rect = new Rect(0, 0, tex.width, tex.height);
-            var pivot = VectorUtils.GetPivot(Alignment, CustomPivot, rect, false);
+
+            // The pivot of the orignial sprite was properly computed, just translate it to the new textured sprite
+            var pivot = sprite.pivot;
+            pivot /= sprite.rect.size;
 
             var texturedSprite = Sprite.Create(tex, rect, pivot, SvgPixelsPerUnit, 0, TexturedSpriteMeshType, m_SpriteData.SpriteRect.border);
             texturedSprite.name = name;